requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.2.0
Flask-Migrate==4.0.7
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
pytest==8.3.5
